third_party/WebKit/LayoutTests/external/wpt/infrastructure/assumptions/html-elements-expected.txt

This is a testharness.js-based test.
PASS (pre-req for comparison tests) all CSS short-hand supported
PASS (pre-req for comparison tests) initial CSS value supported
FAIL Compare CSS div definitions (only valid if pre-reqs pass) assert_equals: Different value for unicode-bidi expected "isolate" but got "normal"
PASS Compare CSS span definitions (only valid if pre-reqs pass)
PASS p is display: block
PASS ul > li is display: list-item
PASS ol > li is display: list-item
PASS table is display: table
PASS tbody is display: table-row-group
PASS tr is display: table-row
PASS td is display: table-cell
Harness: the test ran to completion.
